Move useFullTimeRangeQuery
