+++ web-common/src/features/dashboards/state-managers/loaders/DashboardStateDataLoader.ts
@@ -3,22 +3,25 @@ import {
   type CompoundQueryResult,
   getCompoundQuery,
 } from "@rilldata/web-common/features/compound-query-result";
-import { useFullTimeRangeQuery } from "@rilldata/web-common/features/dashboards/selectors";
 import { cascadingExploreStateMerge } from "@rilldata/web-common/features/dashboards/state-managers/cascading-explore-state-merge";
 import { getPartialExploreStateFromSessionStorage } from "@rilldata/web-common/features/dashboards/state-managers/loaders/explore-web-view-store";
 import { getMostRecentPartialExploreState } from "@rilldata/web-common/features/dashboards/state-managers/loaders/most-recent-explore-state";
 import { getExploreStateFromYAMLConfig } from "@rilldata/web-common/features/dashboards/stores/get-explore-state-from-yaml-config";
 import { getRillDefaultExploreState } from "@rilldata/web-common/features/dashboards/stores/get-rill-default-explore-state";
 import type { MetricsExplorerEntity } from "@rilldata/web-common/features/dashboards/stores/metrics-explorer-entity";
+import { normalizeWeekday } from "@rilldata/web-common/features/dashboards/time-controls/new-time-controls";
 import { convertURLSearchParamsToExploreState } from "@rilldata/web-common/features/dashboards/url-state/convertURLSearchParamsToExploreState";
 import { useExploreValidSpec } from "@rilldata/web-common/features/explores/selectors";
 import {
+  getQueryServiceMetricsViewTimeRangeQueryOptions,
   type V1ExploreSpec,
   type V1MetricsViewSpec,
   type V1MetricsViewTimeRangeResponse,
 } from "@rilldata/web-common/runtime-client";
 import type { AfterNavigate } from "@sveltejs/kit";
-import { get } from "svelte/store";
+import { createQuery, type QueryClient } from "@tanstack/svelte-query";
+import { Settings } from "luxon";
+import { derived, get } from "svelte/store";
 
 /**
  * Loads data from explore and metrics view specs, along with all time range query.
@@ -57,7 +60,7 @@ export class DashboardStateDataLoader {
     private readonly bookmarkOrTokenExploreState?: CompoundQueryResult<Partial<MetricsExplorerEntity> | null>,
   ) {
     this.validSpecQuery = useExploreValidSpec(instanceId, exploreName);
-    this.fullTimeRangeQuery = useFullTimeRangeQuery(
+    this.fullTimeRangeQuery = this.useFullTimeRangeQuery(
       instanceId,
       this.validSpecQuery,
     );
@@ -180,6 +183,92 @@ export class DashboardStateDataLoader {
       rillDefaultExploreState,
       backButtonUsed,
     });
+  }
+
+  /**
+   * Wrapper function that fetches full time range.
+   * Uses useExploreValidSpec unlike the useMetricsViewTimeRange since it is more widely used.
+   *
+   * Does an additional validation where null min and max returned throws an error instead.
+   */
+  private useFullTimeRangeQuery(
+    instanceId: string,
+    validSpecQuery: ReturnType<typeof useExploreValidSpec>,
+    queryClient?: QueryClient,
+  ): CompoundQueryResult<V1MetricsViewTimeRangeResponse> {
+    const fullTimeRangeQueryOptionsStore = derived(
+      validSpecQuery,
+      (validSpecResp) => {
+        const metricsViewSpec = validSpecResp.data?.metricsView ?? {};
+        const exploreSpec = validSpecResp.data?.explore ?? {};
+        const firstDayOfWeek = metricsViewSpec.firstDayOfWeek;
+        const metricsViewName = exploreSpec.metricsView ?? "";
+
+        Settings.defaultWeekSettings = {
+          firstDay: normalizeWeekday(firstDayOfWeek),
+          weekend: [6, 7],
+          minimalDays: 4,
+        };
+
+        return getQueryServiceMetricsViewTimeRangeQueryOptions(
+          instanceId,
+          metricsViewName,
+          {},
+          {
+            query: {
+              enabled: Boolean(metricsViewSpec.timeDimension),
+            },
+          },
+        );
+      },
+    );
+    const fullTimeRangeQuery = createQuery(
+      fullTimeRangeQueryOptionsStore,
+      queryClient,
+    );
+
+    return derived(
+      [fullTimeRangeQueryOptionsStore, fullTimeRangeQuery],
+      ([fullTimeRangeQueryOptions, fullTimeRange]) => {
+        // TODO: update the fields once we move away from getCompoundQuery
+
+        if (!fullTimeRangeQueryOptions.enabled) {
+          // We return early to avoid having isLoading=true when the time range query is not enabled.
+          // This allows us to check isLoading further down without any issues of it getting stuck.
+          // TODO: revisit once we move away from getCompoundQuery
+          return {
+            data: undefined,
+            error: null,
+            isFetching: false,
+            isLoading: false,
+          };
+        }
+
+        if (
+          fullTimeRange.data?.timeRangeSummary?.min === null &&
+          fullTimeRange.data?.timeRangeSummary?.max === null
+        ) {
+          // The timeRangeSummary is null when there are 0 rows of data.
+          // Notably, this happens when a security policy fully restricts a user from reading any data.
+          // Show a different error in this case.
+          return {
+            data: undefined,
+            error: new Error(
+              "This dashboard currently has no data to display. This may be due to access permissions.",
+            ),
+            isFetching: false,
+            isLoading: false,
+          };
+        }
+
+        return {
+          data: fullTimeRange.data,
+          error: fullTimeRange.error,
+          isFetching: fullTimeRange.isFetching,
+          isLoading: fullTimeRange.isLoading,
+        };
+      },
+    );
   }
 
   /**
